Correct abstracted helper method for media search link generation
